# swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Cellnode",
        "version": "1.0.0",
        "description": ""
    },
    "host": "",
    "basePath": "",
    "schemes": ["http"],
    "paths": {
        "/": {
            "get": {
                "description": "",
                "parameters": [],
                "responses": {}
            }
        },
        "/graphql": {
            "post": {
                "description": "GraphQL",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/request": {
            "post": {
                "description": "",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/claims/duplicate": {
            "post": {
                "description": "",
                "parameters": [
                    {
                        "name": "items",
                        "in": "body",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/claims/batch": {
            "post": {
                "description": "",
                "parameters": [
                    {
                        "name": "claims",
                        "in": "body",
                        "required": true,
                        "type": "array"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/storage/store": {
            "post": {
                "description": "",
                "parameters": [
                    {
                        "name": "name",
                        "in": "body",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "contentType",
                        "in": "body",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/storage/retrieve/:cid": {
            "get": {
                "description": "",
                "parameters": [
                    {
                        "name": "cid",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/public/:key": {
            "get": {
                "description": "",
                "parameters": [
                    {
                        "name": "key",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/public/createpublic": {
            "post": {
                "description": "",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/public/fetchpublic": {
            "get": {
                "description": "",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/query/queryransactions": {
            "get": {
                "description": "",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/query/querycapabilities": {
            "get": {
                "description": "",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/capabilities": {
            "post": {
                "description": "",
                "parameters": [
                    {
                        "name": "projectDid",
                        "in": "body",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "userDid",
                        "in": "body",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    }
}